% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_model.R
\name{R6_res_mod}
\alias{R6_res_mod}
\title{R6 Class for antimalarial resistance model}
\description{
Antimalarial resistance model emulator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-res_mod-new}{\code{R6_res_mod$new()}}
\item \href{#method-res_mod-predict}{\code{R6_res_mod$predict()}}
\item \href{#method-res_mod-predict_s}{\code{R6_res_mod$predict_s()}}
\item \href{#method-res_mod-predict_err}{\code{R6_res_mod$predict_err()}}
\item \href{#method-res_mod-predict_t}{\code{R6_res_mod$predict_t()}}
\item \href{#method-res_mod-predict_f2}{\code{R6_res_mod$predict_f2()}}
\item \href{#method-res_mod-add_model}{\code{R6_res_mod$add_model()}}
\item \href{#method-res_mod-add_model_predict_f}{\code{R6_res_mod$add_model_predict_f()}}
\item \href{#method-res_mod-add_model_weight}{\code{R6_res_mod$add_model_weight()}}
\item \href{#method-res_mod-add_err_model}{\code{R6_res_mod$add_err_model()}}
\item \href{#method-res_mod-add_err_model_predict_f}{\code{R6_res_mod$add_err_model_predict_f()}}
\item \href{#method-res_mod-add_err_model_weight}{\code{R6_res_mod$add_err_model_weight()}}
\item \href{#method-res_mod-get_models}{\code{R6_res_mod$get_models()}}
\item \href{#method-res_mod-get_err_models}{\code{R6_res_mod$get_err_models()}}
\item \href{#method-res_mod-get_model_predict_f}{\code{R6_res_mod$get_model_predict_f()}}
\item \href{#method-res_mod-get_err_model_predict_f}{\code{R6_res_mod$get_err_model_predict_f()}}
\item \href{#method-res_mod-get_model_weights}{\code{R6_res_mod$get_model_weights()}}
\item \href{#method-res_mod-get_err_model_weights}{\code{R6_res_mod$get_err_model_weights()}}
\item \href{#method-res_mod-get_data}{\code{R6_res_mod$get_data()}}
\item \href{#method-res_mod-get_test_data}{\code{R6_res_mod$get_test_data()}}
\item \href{#method-res_mod-set_models}{\code{R6_res_mod$set_models()}}
\item \href{#method-res_mod-set_err_models}{\code{R6_res_mod$set_err_models()}}
\item \href{#method-res_mod-set_model_predict_f}{\code{R6_res_mod$set_model_predict_f()}}
\item \href{#method-res_mod-set_err_model_predict_f}{\code{R6_res_mod$set_err_model_predict_f()}}
\item \href{#method-res_mod-set_model_weights}{\code{R6_res_mod$set_model_weights()}}
\item \href{#method-res_mod-set_err_model_weights}{\code{R6_res_mod$set_err_model_weights()}}
\item \href{#method-res_mod-set_data}{\code{R6_res_mod$set_data()}}
\item \href{#method-res_mod-set_test_data}{\code{R6_res_mod$set_test_data()}}
\item \href{#method-res_mod-set_s_adj}{\code{R6_res_mod$set_s_adj()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-new"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-new}{}}}
\subsection{Method \code{new()}}{
Create a new resistance selection model using emulators.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$new(
  models = list(),
  err_models = list(),
  model_weights = list(),
  err_model_weights = list(),
  model_predict_f = list(),
  err_model_predict_f = list(),
  data = data.frame(),
  test_data = data.frame()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{Model for predicting selection coefficient}

\item{\code{err_models}}{Model for predicting error of selection predictions}

\item{\code{model_weights}}{Model for predicting selection coefficient weights}

\item{\code{err_model_weights}}{Model for predicting selection error weights}

\item{\code{model_predict_f}}{Functions to make prediction for specific model}

\item{\code{err_model_predict_f}}{Functions to make prediction for specific error model}

\item{\code{data}}{Data used originally for model training}

\item{\code{test_data}}{Data used originally for model training}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{res_mod} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-predict"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-predict}{}}}
\subsection{Method \code{predict()}}{
Predict median and 95\% selection coefficients and threshold times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$predict(
  al,
  asaq,
  dhappq,
  art_res,
  ppq_res,
  aq_res,
  lu_res,
  ft,
  micro210,
  s_name,
  f1 = 0.01,
  f2 = 0.05,
  sd_scale = 3.3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{al}}{AL drug use}

\item{\code{asaq}}{ASAQ drug use}

\item{\code{dhappq}}{DHAPPQ drug use}

\item{\code{art_res}}{Art-R prevalence}

\item{\code{ppq_res}}{PPQ-R prevalence}

\item{\code{aq_res}}{AQ-R prevalence}

\item{\code{lu_res}}{LU-R prevalence}

\item{\code{ft}}{Treatment Coverage}

\item{\code{micro210}}{Microscopy 2-10 prevalence}

\item{\code{s_name}}{Name of the selection coefficient to be predicted}

\item{\code{f1}}{Starting frequency of mutation}

\item{\code{f2}}{Ending frequency of mutation}

\item{\code{sd_scale}}{Scaling factor for sd based estimation of CI.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.frame of selection coefficients and times from 1\% to 5\%
Predict selection coefficient for data frame of covariates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-predict_s"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-predict_s}{}}}
\subsection{Method \code{predict_s()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$predict_s(dat, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame of covariates}

\item{\code{s_name}}{Name of the selection coefficient to be predicted
Predict error in selection coefficient estimates for covariate data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-predict_err"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-predict_err}{}}}
\subsection{Method \code{predict_err()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$predict_err(dat, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame of covariates}

\item{\code{s_name}}{Name of the selection coefficient to be predicted
Predict t between f1 and f2 and covariate data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-predict_t"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-predict_t}{}}}
\subsection{Method \code{predict_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$predict_t(dat, f1, f2, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame of covariates}

\item{\code{f1}}{Frequency at time point 1}

\item{\code{f2}}{Frequency at time point 2}

\item{\code{s_name}}{Name of the relevant selection coefficient
Predict f2 given f1 and t and covariate data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-predict_f2"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-predict_f2}{}}}
\subsection{Method \code{predict_f2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$predict_f2(dat, f1, t, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame of covariates}

\item{\code{f1}}{Frequency at time point 1}

\item{\code{t}}{Duration of selection}

\item{\code{s_name}}{Name of the relevant selection coefficient
Add models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_model"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_model}{}}}
\subsection{Method \code{add_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_model(model, model_name, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Selection prediction model}

\item{\code{model_name}}{Name of model}

\item{\code{s_name}}{Name of the selection coefficient the model relates to
Add model prediction function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_model_predict_f}{}}}
\subsection{Method \code{add_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_model_predict_f(f, model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{Function to be used for predictions for a specific model}

\item{\code{model_name}}{Name of model
Add model weights}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_model_weight"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_model_weight}{}}}
\subsection{Method \code{add_model_weight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_model_weight(weight, model_name, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weight}}{Model for predicting weights (error) for each model}

\item{\code{model_name}}{Name of model}

\item{\code{s_name}}{Name of the selection coefficient the weights relates to
Add error model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_err_model"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_err_model}{}}}
\subsection{Method \code{add_err_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_err_model(err_model, model_name, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{err_model}}{Selection error prediction model}

\item{\code{model_name}}{Name of model}

\item{\code{s_name}}{Name of the selection coefficient the model relates to
Add model error prediction function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_err_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_err_model_predict_f}{}}}
\subsection{Method \code{add_err_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_err_model_predict_f(f, model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{Function to be used for error predictions for a specific model}

\item{\code{model_name}}{Name of model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-add_err_model_weight"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-add_err_model_weight}{}}}
\subsection{Method \code{add_err_model_weight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$add_err_model_weight(weight, model_name, s_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weight}}{Model for predicting weights (error) for each err model}

\item{\code{model_name}}{Name of model}

\item{\code{s_name}}{Name of the selection coefficient the weights relates to
Get all selection prediction models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_models"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_models}{}}}
\subsection{Method \code{get_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of models
Get all selection error prediction models
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_err_models"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_err_models}{}}}
\subsection{Method \code{get_err_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_err_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of error models
Get all model prediction functions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_model_predict_f}{}}}
\subsection{Method \code{get_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_model_predict_f()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of model prediction functions
Get all error model prediction functions
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_err_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_err_model_predict_f}{}}}
\subsection{Method \code{get_err_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_err_model_predict_f()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of error model prediction functions
Get all selection prediction model weights
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_model_weights}{}}}
\subsection{Method \code{get_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_model_weights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of mode weights
Get all selection error prediction model weights
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_err_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_err_model_weights}{}}}
\subsection{Method \code{get_err_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_err_model_weights()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of error model weights
Get data the models were trained on
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Training data
Get the hold out test data from model training
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-get_test_data"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-get_test_data}{}}}
\subsection{Method \code{get_test_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$get_test_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Training data
Set all selection prediction models
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_models"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_models}{}}}
\subsection{Method \code{set_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_models(models)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{models}}{selection model list
Set all selection error prediction models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_err_models"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_err_models}{}}}
\subsection{Method \code{set_err_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_err_models(err_models)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{err_models}}{selection error model list
Set all model prediction functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_model_predict_f}{}}}
\subsection{Method \code{set_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_model_predict_f(model_predict_f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_predict_f}}{Model prediction function list
Set all model error prediction functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_err_model_predict_f"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_err_model_predict_f}{}}}
\subsection{Method \code{set_err_model_predict_f()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_err_model_predict_f(err_model_predict_f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{err_model_predict_f}}{Model error prediction function list
Set all selection prediction model weights}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_model_weights}{}}}
\subsection{Method \code{set_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_model_weights(model_weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_weights}}{selection model weights (RMSE) list
Set all selection prediction model weights}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_err_model_weights"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_err_model_weights}{}}}
\subsection{Method \code{set_err_model_weights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_err_model_weights(err_model_weights)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{err_model_weights}}{selection error model weights (RMSE) list
Set data the models were trained on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Training data for models
Set hold out test data from model training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_test_data"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_test_data}{}}}
\subsection{Method \code{set_test_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_test_data(test_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_data}}{Training data for models
Set s adjustment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_mod-set_s_adj"></a>}}
\if{latex}{\out{\hypertarget{method-res_mod-set_s_adj}{}}}
\subsection{Method \code{set_s_adj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_mod$set_s_adj(s_adj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s_adj}}{Adjustment for s (NMF fix)}
}
\if{html}{\out{</div>}}
}
}
}
