% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_model.R
\name{R6_res_spread}
\alias{R6_res_spread}
\title{R6 Class for simulating spread of resistance on a given map}
\description{
A simulating model
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-res_spread-new}{\code{R6_res_spread$new()}}
\item \href{#method-res_spread-set_seeds}{\code{R6_res_spread$set_seeds()}}
\item \href{#method-res_spread-set_map_data}{\code{R6_res_spread$set_map_data()}}
\item \href{#method-res_spread-simulate_spread}{\code{R6_res_spread$simulate_spread()}}
\item \href{#method-res_spread-simulate_multiallelic_spread}{\code{R6_res_spread$simulate_multiallelic_spread()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_spread-new"></a>}}
\if{latex}{\out{\hypertarget{method-res_spread-new}{}}}
\subsection{Method \code{new()}}{
Create a new res spread model object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_spread$new(map, res_mod, adj_mat = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{Map sf object}

\item{\code{res_mod}}{res_mod selection model}

\item{\code{adj_mat}}{Adjacency matrix for \code{map}. Default = NULL and will calculate internally}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_spread-set_seeds"></a>}}
\if{latex}{\out{\hypertarget{method-res_spread-set_seeds}{}}}
\subsection{Method \code{set_seeds()}}{
Set up simulation seeds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_spread$set_seeds(seeds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seeds}}{Named vector or list of regions seeding spread and iniial frequency
E.g. list("region_1" = 0.2, "region_2" = 0.3, "region_3" = 0.002)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_spread-set_map_data"></a>}}
\if{latex}{\out{\hypertarget{method-res_spread-set_map_data}{}}}
\subsection{Method \code{set_map_data()}}{
Set up map_data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_spread$set_map_data(map_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map_data}}{Data.frame with region names ("id_1") and selection ("s")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_spread-simulate_spread"></a>}}
\if{latex}{\out{\hypertarget{method-res_spread-simulate_spread}{}}}
\subsection{Method \code{simulate_spread()}}{
Simulated spread
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_spread$simulate_spread(
  import_freq = 0.01,
  export_freq = 0.25,
  t_end = 40,
  t_break = 1,
  import_gap = 1,
  s_name = "s_a_5"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{import_freq}}{What frequency does importation result in. Default = 0.01}

\item{\code{export_freq}}{At what frequency does exportation occur at. Default = 0.25}

\item{\code{t_end}}{What year does simulation end. Default = 40}

\item{\code{t_break}}{Gap between time breaks. Default = 1}

\item{\code{import_gap}}{Number of years for importation to occur over. Default = 1}

\item{\code{s_name}}{Name of selection coefficient to be simulated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res_spread-simulate_multiallelic_spread"></a>}}
\if{latex}{\out{\hypertarget{method-res_spread-simulate_multiallelic_spread}{}}}
\subsection{Method \code{simulate_multiallelic_spread()}}{
Simulated spread of multiallelic traits
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_res_spread$simulate_multiallelic_spread(
  import_freq = 0.01,
  export_freq = 0.25,
  t_end = 40,
  t_break = 1,
  import_gap = 1,
  s_names = paste0("s_a_", 1:6)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{import_freq}}{What frequency does importation result in. Default = 0.01}

\item{\code{export_freq}}{At what frequency does exportation occur at. Default = 0.25}

\item{\code{t_end}}{What year does simulation end. Default = 40}

\item{\code{t_break}}{Gap between time breaks. Default = 1}

\item{\code{import_gap}}{Number of years for importation to occur over. Default = 1}

\item{\code{s_names}}{Names of selection coefficient to be simulated}
}
\if{html}{\out{</div>}}
}
}
}
